streamlit/requirements.txt
streamlit==1.29.0
pandas==2.1.3
psycopg2-binary==2.9.9
plotly==5.18.0